Minimal constraint fix
The storyboard had a center y minus a fixed amount - I changed this to a
fixed heigh for the view as I think that was (roughly) the intention - the
other way caused weirdness of different device size. I think we need to
go back and respecify some of these constraints, I will take a look next week.

Change-Id: Ia3e162d80b1128ad11bf4d42270030319ef53f96